feat (refactor): 重构项目结构
