+++ src/ops/c_interface/cuda/nv_gpu.cu
@@ -1,19 +1,23 @@
-﻿#include "cpu.h"
-#include "rms_norm.h"
+﻿#include "nv_gpu.cuh"
+#include "../../rms_norm/cuda/rms_norm.cuh"
+#include "../../../utils.h"
 #include <cstdio>
 
 static void kn_drop(Kn kn) {
+    delete reinterpret_cast<NvGpuRtCtx *>(kn->rt_ctx);
     delete kn;
 }
 
-static Kn load(Op op, void *) {
+static Kn load(Op op, void *rt_ctx) {
+    ASSERT_VALID_PTR(rt_ctx);
+    auto ctx = new NvGpuRtCtx(*reinterpret_cast<NvGpuRtCtx *>(rt_ctx));
     switch (op->optype) {
         case OpRmsNorm: {
             auto kn = new Kernel{
-                DevCpu,
+                DevNvGpu,
                 OpRmsNorm,
-                nullptr,
-                (Fn) rms_norm_cpu_f16,
+                ctx,
+               (Fn) rms_norm_nv_gpu_f16,
                 kn_drop,
             };
             return kn;
@@ -37,11 +41,11 @@ static void op_drop(Op op) {
     delete op;
 }
 
-Op op_create_cpu(Optype opty, void *) {
+Op op_create_nv_gpu(Optype opty, void *) {
     switch (opty) {
         case OpRmsNorm: {
             auto op = new Operator{
-                DevCpu,
+                DevNvGpu,
                 OpRmsNorm,
                 nullptr,
                 load,
