+++ src/ops/rms_norm/cpu/rms_norm.h
@@ -1,7 +1,7 @@
 ﻿#ifndef __CPU_RMS_NORM_H__
 #define __CPU_RMS_NORM_H__
 
-#include "../operators.h"
+#include "../../../operators.h"
 
 void rms_norm_cpu_f16(Kernel const *, MutTensor y, ConstTensor x, ConstTensor w, float epsilon);
