+++ src/ops/rms_norm/cuda/rms_norm.cuh
@@ -1,7 +1,7 @@
 ﻿#ifndef __NV_GPU_RMS_NORM_H__
 #define __NV_GPU_RMS_NORM_H__
 
-#include "../operators.h"
+#include "../../../operators.h"
 
 void rms_norm_nv_gpu_f16(Kernel const *, MutTensor y, ConstTensor x, ConstTensor w, float epsilon);
