+++ xmake.lua
@@ -14,13 +14,15 @@ option("nv-gpu")
     add_defines("ENABLE_NV_GPU")
 option_end()
 
+if is_mode("debug") then
+    add_cxflags("-g -O0")
+    add_defines("DEBUG_MODE")
+end
+
 target("common")
     set_kind("static")
     set_languages("c11")
     add_files("src/utils.c")
-    if is_mode("debug") then
-        add_cxflags("-g")
-    end
 target_end()
 
 if has_config("cpu") then
@@ -28,11 +30,8 @@ add_defines("ENABLE_CPU")
 target("cpu")
     set_kind("static")
     set_languages("cxx17")
-    add_files("src/cpu/*.cc")
+    add_files("src/ops/*/cpu/*.cc")
     add_deps("common")
-    if is_mode("debug") then
-        add_cxflags("-g")
-    end
 target_end()
 
 end
@@ -43,14 +42,11 @@ add_defines("ENABLE_NV_GPU")
 target("nv-gpu")
     set_kind("static")
     set_languages("cxx17")
-    add_cuflags("-arch=sm_80", "--expt-relaxed-constexpr", "--allow-unsupported-compiler")
-    add_files("src/nv_gpu/*.cu")
+    add_cuflags("-arch=sm_80", "--expt-relaxed-constexpr", "--allow-unsupported-compiler", {force = true})
+    add_files("src/ops/*/cuda/*.cu")
     set_toolchains("cuda")
     set_policy("build.cuda.devlink", true)
     add_deps("common")
-    if is_mode("debug") then
-        add_cxflags("-g")
-    end
 target_end()
 
 end
@@ -66,9 +62,6 @@ end
 if has_config("nv-gpu") then
     add_deps("nv-gpu")
 end
-if is_mode("debug") then
-    add_cxflags("-g")
-end
 target_end()
 
 target("main")
@@ -76,7 +69,4 @@ target("main")
     set_languages("cxx17")
     add_files("src/main.cpp")
     add_deps("operators")
-    if is_mode("debug") then
-        add_cxflags("-g")
-    end
 target_end()
